requirements.txt
Django>=4.2.5,<4.3
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.26.5,<0.27
mysqlclient>=2.0.3,<2.1
django-cors-headers==4.3.0
pandas==2.1.2
redis>=3.5
celery>=5.0